+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-24 11:38:53
-// Commit: fix(test): fix memory leak by cancelling test subscription
+// Updated at: 2026-08-24 20:58:56
+// Commit: refactor(test): optimize data flow between domain and data layers
 
 class testComponentState {
-  final String commitRef = '3877';
-  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
-  final DateTime timestamp = DateTime.parse('2026-08-24T11:38:53+05:00');
+  final String commitRef = '3915';
+  final String lastMessage = 'refactor(test): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-08-24T20:58:56+05:00');
 }